# server/src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: GAIA Cosmetics API
  description: Backend API for GAIA Cosmetics E-commerce Website
  version: 1.0.0
  contact:
    name: API Support
    email: support@gaiacosmetics.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.gaiacosmetics.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
    ValidationError:
      type: object
      properties:
        message:
          type: string
          description: Validation error message
        errors:
          type: array
          items:
            type: string
            description: Detailed error messages
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        original_price:
          type: number
          format: float
          nullable: true
        discount_percentage:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        ingredients:
          type: string
          nullable: true
        how_to_use:
          type: string
          nullable: true
        inventory_status:
          type: string
          enum: [in-stock, low-stock, out-of-stock]
        inventory_message:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              imagePath:
                type: string
              webpPath:
                type: string
                description: WebP version of the image path (for better performance)
              isPrimary:
                type: boolean
        colors:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              value:
                type: string
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image_path:
          type: string
          nullable: true
        webp_path:
          type: string
          nullable: true
          description: WebP version of the image path
        product_count:
          type: integer
    Content:
      type: object
      properties:
        id:
          type: integer
        section:
          type: string
        key:
          type: string
        value:
          type: string
          nullable: true
        type:
          type: string
          enum: [text, image, html, video]
        updated_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get all products with pagination
      security: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category
        - in: query
          name: search
          schema:
            type: string
          description: Search term for product name or description
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  pagination:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                      totalPages:
                        type: integer
                      totalItems:
                        type: integer
                      perPage:
                        type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '429':
          description: Too many requests, please try again later
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Products
      summary: Create a new product (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - category
                - price
              properties:
                name:
                  type: string
                  example: Luminous Foundation
                category:
                  type: string
                  example: Makeup
                price:
                  type: number
                  format: float
                  example: 29.99
                originalPrice:
                  type: number
                  format: float
                  example: 39.99
                discountPercentage:
                  type: integer
                  example: 25
                description:
                  type: string
                  example: A lightweight foundation for all skin types
                ingredients:
                  type: string
                  example: Water, Cyclopentasiloxane, Titanium Dioxide
                howToUse:
                  type: string
                  example: Apply a small amount to clean skin
                inventoryStatus:
                  type: string
                  example: in-stock
                inventoryMessage:
                  type: string
                  example: Available for immediate shipping
                colors:
                  type: string
                  example: '[{"name":"Light","value":"#F5DEB3"},{"name":"Medium","value":"#D2B48C"}]'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  message:
                    type: string
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/content:
    get:
      tags:
        - Content
      summary: Get all website content
      security: []
      responses:
        '200':
          description: Website content grouped by section
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties:
                    type: object
                    properties:
                      value:
                        type: string
                      type:
                        type: string
                      id:
                        type: integer
        '429':
          description: Too many requests, please try again later
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Content
      summary: Add new content (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - section
                - key
                - value
              properties:
                section:
                  type: string
                  example: hero
                key:
                  type: string
                  example: title
                value:
                  type: string
                  example: Ready to look flawless all day
                type:
                  type: string
                  enum: [text, image, html, video]
                  default: text
                  example: text
      responses:
        '201':
          description: Content added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          description: Invalid input or content already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      security: []
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '429':
          description: Too many requests, please try again later
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Categories
      summary: Create a new category (Admin only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Skincare
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Category name is required or category already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...